Made progress on connecting back-end API to front-end, just completed Edit and moving onto Add New.

+++ server/controllers/userController.js
@@ -13,6 +13,19 @@ exports.index = (req, res) => {
     });
 };
 
+exports.getSingleUser = (req, res) => {
+  knex("user")
+    .where({ userID: req.params.userID })
+    .then((data) => {
+      res.json(data);
+    })
+    .catch((err) => {
+      res
+        .status(400)
+        .send(`Ran into an error while trying to get the user info: ${err}`);
+    });
+};
+
 exports.userMemories = (req, res) => {
   knex("memory")
     .where({ userID: req.params.userID })
@@ -25,6 +38,19 @@ exports.userMemories = (req, res) => {
         .send(
           `Ran into an error while trying to get the user's memories: ${err}`
         );
+    });
+};
+
+exports.getMemoryByUser = (req, res) => {
+  knex("memory")
+    .where({ userID: req.params.userID, memoryID: req.params.memoryID })
+    .then((data) => {
+      res.json(data);
+    })
+    .catch((err) => {
+      res
+        .status(400)
+        .send(`Ran into an error while trying to get the memory info: ${err}`);
     });
 };
 
@@ -49,7 +75,7 @@ exports.addMemoryByUser = (req, res) => {
 exports.editMemoryByUser = (req, res) => {
   knex("memory")
     .update(req.body)
-    .where({ memoryID: req.params.memoryID })
+    .where({ userID: req.params.userID, memoryID: req.params.memoryID })
     .then((data) => {
       res.json({
         message: "This memory has been updated successfully",
@@ -66,7 +92,7 @@ exports.editMemoryByUser = (req, res) => {
 exports.deleteMemoryByUser = (req, res) => {
   knex("memory")
     .delete()
-    .where({ memoryID: req.params.memoryID })
+    .where({ userID: req.params.userID, memoryID: req.params.memoryID })
     .then((data) => {
       res.status(204).json({
         message: "This memory has been deleted successfully",
